+++ lib/geordi/commands/rspec.rb
@@ -9,7 +9,7 @@ LONGDESC
 def rspec(*files)
   if File.exists?('spec/spec_helper.rb')
     invoke_cmd 'bundle_install'
-    invoke_cmd 'yarn'
+    invoke_cmd 'yarn_install'
 
     announce 'Running specs'