+++ lib/geordi/commands/unit.rb
@@ -2,7 +2,7 @@ desc 'unit', 'Run Test::Unit'
 def unit
   if File.exists?('test/test_helper.rb')
     invoke_cmd 'bundle_install'
-    invoke_cmd 'yarn'
+    invoke_cmd 'yarn_install'
 
     announce 'Running Test::Unit'
     Util.system! 'bundle exec rake test'