+++ lib/geordi/commands/with_rake.rb
@@ -2,7 +2,7 @@ desc 'with-rake', 'Run tests with `rake`', :hide => true
 def with_rake
   if Util.file_containing?('Rakefile', /^task.+default.+(spec|test|feature)/)
     invoke_cmd 'bundle_install'
-    invoke_cmd 'yarn'
+    invoke_cmd 'yarn_install'
 
     announce 'Running tests with `rake`'
     Util.system! 'rake'